@@ -206,33 +216,33 @@ val compileWasm = task<Exec>("compileWasm") {
 
 fun path(more: Iterable<*>) = more.joinToString(File.separator)
 
-val genParser = task<GenerateParser>("genParser") {
+val genParser = task<GenerateParserTask>("genParser") {
 	group = tasks["init"].group!!
 	description = "Generate the Parser and PsiElement classes"
-	source = "grammar/pest.bnf"
-	targetRoot = "gen/"
+	source.value("grammar/pest.bnf")
+	targetRoot.value("gen/")
 	val parserRoot = Paths.get("rs", "pest")
-	pathToParser = path(parserRoot + "PestParser.java")
-	pathToPsiRoot = path(parserRoot + "psi")
-	purgeOldFiles = true
+	pathToParser.value(path(parserRoot + "PestParser.java"))
+	pathToPsiRoot.value(path(parserRoot + "psi"))
+	purgeOldFiles.value(true)
 }
 
-val genLexer = task<GenerateLexer>("genLexer") {
+val genLexer = task<GenerateLexerTask>("genLexer") {
 	group = genParser.group
 	description = "Generate the Lexer"
-	source = "grammar/pest.flex"
-	targetDir = path(Paths.get("gen", "rs", "pest", "psi"))
-	targetClass = "PestLexer"
-	purgeOldFiles = true
+	source.value("grammar/pest.flex")
+	targetDir.value(path(Paths.get("gen", "rs", "pest", "psi")))
+	targetClass.value("PestLexer")
+	purgeOldFiles.value(true)
 	dependsOn(genParser)
 }
 
 tasks.withType<KotlinCompile> {
 	dependsOn(genParser, genLexer, compileWasm)
 	kotlinOptions {
-		jvmTarget = "1.8"
-		languageVersion = "1.3"
-		apiVersion = "1.3"
+		jvmTarget = "11"
+		languageVersion = "1.6"
+		apiVersion = "1.6"
 		freeCompilerArgs = listOf("-Xjvm-default=enable")
 	}
 }
